refactor: move authStateFn to a new file and update import paths
